moved eval related code into the utils file
